GopiAI-CrewAI/requirements.txt
# 🚀 GopiAI-CrewAI Integration Requirements
# Полный список зависимостей для интеграции CrewAI с GopiAI

# === ОСНОВНЫЕ ЗАВИСИМОСТИ ===
crewai==0.130.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.0.0

# === LLM ПРОВАЙДЕРЫ ===
# Groq LLM (быстрые модели)
langchain-groq>=0.1.0

# Google Gemini (стабильные модели) 
langchain-google-genai>=1.0.0

# OpenAI совместимые (Cerebras, Novita и др.)
langchain-openai>=0.1.0
openai>=1.0.0

# Cohere LLM
cohere>=5.0.0

# Hugging Face LLM (добавлено для интеграции)
transformers>=4.30.0
huggingface-hub>=0.16.0

# === ДОПОЛНИТЕЛЬНЫЕ ИНСТРУМЕНТЫ ===
# Для браузера и веб-запросов  
beautifulsoup4>=4.12.0
lxml>=4.9.0
selenium>=4.15.0  # Опционально для сложной автоматизации

# Для работы с файлами
chardet>=5.2.0  # Определение кодировки файлов
pathlib2>=2.3.0  # Расширенная работа с путями

# Для памяти и векторного поиска
numpy>=1.24.0
scikit-learn>=1.3.0  # Для семантического поиска
sentence-transformers>=2.2.0  # Опционально для embeddings

# === ИНТЕГРАЦИЯ С GOPIAI ===
# Asyncio для асинхронной коммуникации
aiohttp>=3.8.0
websockets>=11.0.0  # Для real-time коммуникации с UI

# JSON обработка и сериализация
ujson>=5.8.0  # Быстрый JSON
jsonschema>=4.19.0  # Валидация схем

# === УТИЛИТЫ И ДИАГНОСТИКА ===
# Логирование и мониторинг
colorlog>=6.7.0  # Цветные логи
rich>=13.5.0  # Красивый вывод в консоль
tqdm>=4.66.0  # Прогресс-бары

# Тестирование
pytest>=7.4.0
pytest-asyncio>=0.21.0

# === ОПЦИОНАЛЬНЫЕ ЗАВИСИМОСТИ ===
# Раскомментируйте при необходимости

# Для расширенной работы с изображениями
# Pillow>=10.0.0
# opencv-python>=4.8.0

# Для работы с базами данных
# sqlalchemy>=2.0.0
# sqlite3  # Встроен в Python

# Для ML и нейросетей
# torch>=2.0.0
# transformers>=4.30.0

# Для работы с PDF и документами
# PyPDF2>=3.0.0
# python-docx>=0.8.11

# Для продвинутой веб-автоматизации
# playwright>=1.37.0
# httpx>=0.24.0

# === СИСТЕМНЫЕ ТРЕБОВАНИЯ ===
# Python >= 3.8
# RAM >= 4GB (рекомендуется 8GB)
# Интернет соединение для LLM API
# Место на диске >= 1GB для кеша и памяти